+++ chartpress.py
@@ -336,27 +336,49 @@ def build_values(name, values_mods):
         values = yaml.load(f)
 
     for key, value in values_mods.items():
+        if not isinstance(value, dict) or set(value.keys()) != {'repository', 'tag'}:
+            raise ValueError(f"I only understand image updates with 'repository', 'tag', not: {value!r}")
         parts = key.split('.')
-        mod_obj = values
+        mod_obj = parent = values
         for p in parts:
+            if p.isdigit():
+                # integers are indices in lists
+                p = int(p)
+            parent = mod_obj
             mod_obj = mod_obj[p]
-        print(f"Updating {values_file}: {key}: {value}")
+            last_part = p
 
         if isinstance(mod_obj, MutableMapping):
             keys = IMAGE_REPOSITORY_KEYS & mod_obj.keys()
             if keys:
-                for key in keys:
-                    mod_obj[key] = value['repository']
+                for repo_key in keys:
+                    before = mod_obj.get(repo_key, None)
+                    if before != value['repository']:
+                        print(f"Updating {values_file}: {key}.{repo_key}: {value}")
+                    mod_obj[repo_key] = value['repository']
             else:
                 possible_keys = ' or '.join(IMAGE_REPOSITORY_KEYS)
                 raise KeyError(
                     f'Could not find {possible_keys} in {values_file}:{key}'
                 )
 
+            before = mod_obj.get('tag', None)
+            if before != value['tag']:
+                print(f"Updating {values_file}: {key}.tag: {value}")
             mod_obj['tag'] = value['tag']
+        elif isinstance(mod_obj, str):
+            # scalar image string, not dict with separate repository, tag keys
+            image = "{repository}:{tag}".format(**value)
+            try:
+                before = parent[last_part]
+            except (KeyError, IndexError):
+                before = None
+            if before != image:
+                print(f"Updating {values_file}: {key}: {image}")
+            parent[last_part] = image
         else:
             raise TypeError(
-                f'The key {key} in {values_file} must be a mapping.'
+                f'The key {key} in {values_file} must be a mapping or string, not {type(mod_obj)}.'
             )